backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pymongo==4.3.3
python-dotenv==1.0.0
gspread==5.12.0
google-auth==2.25.2
pandas==2.1.3
pytest
